msm: allow to use the phone as a bootable USB key
With the current code, my BIOS fails to boot from the USB Mass storage
due the fact that it sends a CDB INQUIRY with bCBWCBLength==12 while the INQUIRY
command size is equal to 6.

According to USB Mass Storage Specification for Bootability,
( http://www.usb.org/developers/devclass_docs/usb_msc_boot_1.0.pdf )
"The device shall accept a BOT bCBWCBLength of 12 (0Ch) for all the commands
documented in this specification." and "PAD bytes are added to make the CDBs
have a fixed length of 12 (0Ch). Pad bytes shall be set to zero by the host."

So, let check_command() also accept a fixed size of 12 and then check that
the remaining bytes are really 0 (by updating cmnd_size).

Tested successfully with several different BIOS.

Signed-off-by: Vincent Palatin <vincent.palatin_android@m4x.org>